Return the more concise intersection test algorithm
This more readable algorithm is as fast as the other non-readable one, although this one
uses construction of new arrays and function calls along the way, and the other does it all inline.
C++ seemingly can compile all that to the same or similar machine code. THis is not the case in Java, where
new Point() somewhere in the middle of the algorithm led to a significant drop of speed.